Pre-mod cleanup, DB generation file
