+++ guessdict.log
